+++ apps/web/src/store/draft.js
@@ -3,6 +3,5 @@ import { defineStore } from 'pinia'
 export const userDraftStore = defineStore('draft', {
     state: () => ({
         refreshList: 0,
-        setDraft: {},
     })
 })